requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
kafka-python==2.0.2
requests==2.26.0
tkinter 